+++ spec/lib/parameter_substitution_spec.rb
@@ -81,6 +81,18 @@ describe ParameterSubstitution do
         it "returns tokens that exist in mapping when one is provided" do
           expect(ParameterSubstitution.find_tokens(expression, mapping: mapping)).to eq(['call.start_time', 'do_a_barrel_roll'])
         end
+
+        context 'with non-default delimiters' do
+          let(:square_expression) { "[call.start_time.blank_if_nil][do_a_barrel_roll.downcase]" }
+
+          it "returns tokens up to first dot when no mapping is provided" do
+            expect(ParameterSubstitution.find_tokens(square_expression, parameter_start: "[", parameter_end: "]")).to eq(['call', 'do_a_barrel_roll'])
+          end
+
+          it "returns tokens that exist in mapping when one is provided" do
+            expect(ParameterSubstitution.find_tokens(square_expression, mapping: mapping, parameter_start: "[", parameter_end: "]")).to eq(['call.start_time', 'do_a_barrel_roll'])
+          end
+        end
       end
 
       context '#find_formatters' do
@@ -90,6 +102,18 @@ describe ParameterSubstitution do
 
         it "returns formatters after all tokens when mapping is provided" do
           expect(ParameterSubstitution.find_formatters(expression, mapping: mapping)).to eq(['blank_if_nil', 'downcase'])
+        end
+
+        context 'with non-default delimiters' do
+          let(:square_expression) { "[call.start_time.blank_if_nil][do_a_barrel_roll.downcase]" }
+
+          it "returns all formatters after first dot when no mapping is provided" do
+            expect(ParameterSubstitution.find_formatters(square_expression, parameter_start: "[", parameter_end: "]")).to eq(['start_time', 'blank_if_nil', 'downcase'])
+          end
+
+          it "returns formatters after all tokens when mapping is provided" do
+            expect(ParameterSubstitution.find_formatters(square_expression, mapping: mapping, parameter_start: "[", parameter_end: "]")).to eq(['blank_if_nil', 'downcase'])
+          end
         end
       end
 
@@ -125,6 +149,28 @@ describe ParameterSubstitution do
           it "returns 2 warnings" do
             expect(ParameterSubstitution.find_warnings(expression_with_bad_params_and_methods))
               .to eq(["Unknown param 'bobby' and methods 'test1', 'test2'", "Unknown param 'bobby2' and methods 'test3', 'test4'"])
+          end
+        end
+
+        context "with non-default delimiters" do
+          let(:square_expression_with_valid_params) { "[foo]" }
+          let(:square_expression_with_bad_params) { "[bobby][bobby2]" }
+          let(:square_expression_with_bad_methods) { "[foo.test1.test2][foo.test3.test4][black.test1.test2]" }
+
+          it "returns empty array for valid parameters" do
+            expect(ParameterSubstitution.find_warnings(square_expression_with_valid_params, mapping: default_mapping, parameter_start: "[", parameter_end: "]"))
+              .to eq([])
+          end
+
+          it "returns warnings for invalid parameters" do
+            expect(ParameterSubstitution.find_warnings(square_expression_with_bad_params, parameter_start: "[", parameter_end: "]"))
+              .to eq(["Unknown param 'bobby'", "Unknown param 'bobby2'"])
+          end
+
+          it "returns warnings for invalid methods" do
+            expect(ParameterSubstitution.find_warnings(square_expression_with_bad_methods, mapping: default_mapping, parameter_start: "[", parameter_end: "]"))
+              .to eq(["Unknown methods 'test1', 'test2', 'test3', 'test4' used on parameter 'foo'",
+                      "Unknown methods 'test1', 'test2' used on parameter 'black'"])
           end
         end
       end