+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Inspired by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Note: This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2025-11-06
+### Added
+- Added optional `parameter_start` and `parameter_end` arguments for the following methods:
+  - `ParameterSubstitution.find_tokens`
+  - `ParameterSubstitution.find_formatters`
+  - `ParameterSubstitution.find_warnings`
+
 ## [2.0.0] - 2024-07-11
 ### Removed
 - Remove support for eol ruby versions 2.7